be-e3/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
python-dotenv
retell-sdk
openai
requests
